+++ src/gui/UIConst.ts
@@ -2,4 +2,5 @@ export class UIConst {
         public static _diagWidth: number = 480;
         public static _diagWidthS: string = "28em";
         public static _buttonHeight: number = 41;
+        public static MENU_BAR_HEIGHT: number = 48;
 }